% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.athena-package.R
\docType{package}
\name{aws.athena}
\alias{aws.athena}
\alias{aws.athena-package}
\title{Access Amazon's AWS Athena API}
\description{
Reticulated wrapper for the Python 'boto3' Athena client library.
}
\note{
This package \strong{requires} Python >= 3.5 to be available and the \code{boto3} Python package.
}
\author{
Bob Rudis (bob@rud.is)
}
