% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-named-query.R
\name{create_named_query}
\alias{create_named_query}
\title{Create a named query.}
\usage{
create_named_query(name, description, query, database,
  client_request_token = uuid::UUIDgenerate(), encryption_option = NULL,
  kms_key = NULL, aws_access_key_id = NULL, aws_secret_access_key = NULL,
  aws_session_token = NULL, region_name = NULL, profile_name = NULL)
}
\arguments{
\item{name}{plain language name for the query}

\item{description}{brief explanation of the query}

\item{query}{text of the query itself. In other words, all query statements}

\item{database}{database to which the query belongs}

\item{client_request_token}{unique case-sensitive string used to ensure the
request to create the query is idempotent (executes only once). If another
\code{CreateNamedQuery} request is received, the same response is returned
and another query is not created. If a parameter has changed, for example,
the \code{query} , an error is returned. \strong{This is auto-generated-for-you}.}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Create a named query.
}
\examples{
\dontrun{
create_named_query(
  name = "elb100",
  description = "100 rows from elb_logs",
  query = "SELECT * FROM elb_logs LIMIT 100",
  database = "sampledb"
)
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.create_named_query}
}
