% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-named-query.R
\name{delete_named_query}
\alias{delete_named_query}
\title{Delete a named query.}
\usage{
delete_named_query(named_query_id, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{named_query_id}{unique ID of the query to delete}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Delete a named query.
}
\examples{
\dontrun{
res <- list_named_queries()

delete_named_query(res$NamedQueryIds[1])
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.create_named_query}
}
