% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-and-save-query.R
\name{execute_and_save_query}
\alias{execute_and_save_query}
\title{Execute a Query and Save it to disk}
\usage{
execute_and_save_query(query, database, output_location, local_filename,
  client_request_token = NULL, encryption_option = NULL,
  kms_key = NULL, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL, polling_duration = 15,
  checktime = FALSE, maxtime = 1000)
}
\arguments{
\item{query}{SQL query statements to be executed}

\item{database}{database within which the query executes.}

\item{output_location}{location in S3 where query results are stored.}

\item{local_filename}{download location of the athena result}

\item{client_request_token}{unique case-sensitive string used to ensure the
request to create the query is idempotent (executes only once). If another
\code{StartQueryExecution} request is received, the same response is returned
and another query is not created. If a parameter has changed, for example,
the \code{query} , an error is returned. \strong{This is auto-generated-for-you}.}

\item{encryption_option}{indicates whether Amazon S3 server-side encryption
with Amazon S3-managed keys (\code{SSE-S3}), server-side encryption with
KMS-managed keys (\code{SSE-KMS}), or client-side encryption with KMS-managed
keys (\code{CSE-KMS}) is used. Default is \code{NULL} (no encryption)}

\item{kms_key}{For \code{SSE-KMS} and \code{CSE-KMS}, this is the KMS key ARN or ID.
Default is \code{NULL} (no encryption)}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{polling_duration}{how often to check for results}

\item{checktime}{whether to check for time}

\item{profile}{profile name}

\item{max_time}{how long to keep looking for data}
}
\value{
the filename
}
\description{
This funciton combines the \code{start_query_execution} with the
\code{save_query_results} function to execute, run, and retrieve
the results of a SQL query on Athena.
}
\details{
This is useful for downloading large results where it is undesireable
to load the full dataset into memory.
}
\examples{
\dontrun{
execute_and_save_query(
  query    = "SELECT * FROM elb_logs LIMIT 100",
  database = "sampledb",
  profile  = "personal",
  output_location = "s3://aws-athena-query-results-redacted",
  local_filename = "download-to-here.csv",
  polling_duration = 15,
  checktime=FALSE,
  maxtime=1000
  )
}


}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.get_query_results}
}
