% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-batch-named-queries.R
\name{get_named_queries}
\alias{get_named_queries}
\title{Get Query Execution (batch/multiple)}
\usage{
get_named_queries(named_query_ids, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{named_query_ids}{character vector of 1-50 unique named query ids}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Returns information about a single query.
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.batch_get_named_query}
}
