% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-named-query.R
\name{get_named_query}
\alias{get_named_query}
\title{Get Query Execution}
\usage{
get_named_query(named_query_id, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{named_query_id}{unique ID of the named query}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Returns information about a single query.
}
\examples{
\dontrun{
res <- list_named_queries()

get_named_query(res$NamedQueryIds[1])
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.get_named_query}
}
