% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-query-executions.R
\name{get_query_executions}
\alias{get_query_executions}
\title{Get Query Executions (batch/multiple)}
\usage{
get_query_executions(query_execution_ids, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{query_execution_ids}{character vector of 1-50 unique query execution IDs.}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Returns the details of a single query execution or a list of up to 50 query
executions, which you provide as an array of query execution ID strings. To
get a list of query execution IDs, use \code{\link[=list_query_executions]{list_query_executions()}}. Query
executions are different from named (saved) queries. Use
\code{\link[=batch_get_named_queries]{batch_get_named_queries()}} to get details about named queries.
}
\examples{
\dontrun{
res <- list_query_executions()

get_query_executions(res$QueryExecutionIds)
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.batch_get_query_execution}
}
