% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-query-results.R
\name{get_query_results}
\alias{get_query_results}
\title{Get Query Results}
\usage{
get_query_results(query_execution_id, chunk_size = 1000L,
  aws_access_key_id = NULL, aws_secret_access_key = NULL,
  aws_session_token = NULL, region_name = NULL, profile_name = NULL)
}
\arguments{
\item{query_execution_id}{unique ID of the query execution.}

\item{chunk_size}{the AWS Athena API returns the result set in batches. Smaller
values for \code{chunk_size} mean more requests are made to retrieve the
entire result set. Larger values for \code{chunk_size} will retrieve more
data for each call but may not be as performant depending on many
factors including network speed, AWS region, etc. Tune this if the
default causes API failures for your situation. The maximum value
(set by the AWS Athena API endpoint) is 1000.}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\value{
data frame (tibble)
}
\description{
Returns the results of a single query execution specified by
\code{query_execution_id}. This request does not execute the query but returns
results. Use \code{\link[=start_query_execution]{start_query_execution()}} to run a query.
}
\note{
This retrieves \emph{all} the results (i.e. it handles pagination for you).
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.get_query_results}
}
