% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-named-queries.R
\name{list_named_queries}
\alias{list_named_queries}
\title{List Named Queries}
\usage{
list_named_queries(max_results = 50L, next_token = NULL,
  aws_access_key_id = NULL, aws_secret_access_key = NULL,
  aws_session_token = NULL, region_name = NULL, profile_name = NULL)
}
\arguments{
\item{max_results}{maximum number of query executions to return in this request}

\item{next_token}{token that specifies where to start pagination if a previous
request was truncated.}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Provides a list of all available query IDs.
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.list_named_queries}
}
