% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-query-execution.R
\name{stop_query_execution}
\alias{stop_query_execution}
\title{Stop Query Execution}
\usage{
stop_query_execution(query_execution_id, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{query_execution_id}{unique ID of the query execution.}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Stops a query execution.
}
\examples{
\dontrun{
start_query_execution(
  query = "SELECT * FROM elb_logs LIMIT 100",
  database = "sampledb",
  output_location = "s3://aws-athena-query-results-redacted",
  profile = "personal"
) -> sqe

stop_query_execution(sqe)
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/athena.html#Athena.Client.stop_query_execution}
}
